Allow nodes to be dynamically added and removed
This was done by manually adding new nodes and deleting old ones - d3 recognizes deleting a node
and adding an identical one as adding a new node. This causes nodes to re-appear in the corner and is jarring.
Thus, it was necessary to only change nodes to be added and deleted.

Unfortunately, there was no easy way to only update affected nodes and links. The simple solution was to
update all nodes and links on any database change.

Please note that all links attached to a node must be deleted before the node can be deleted.